refactor(routing and wording): move vote to proposal for all routes and wordings
