+++ src/pages/proposal/active/GroupHeader.tsx
