+++ src/pages/proposal/completed/NewMemberVote.tsx
