+++ src/pages/proposal/create/NoVotingType.tsx
