+++ src/routes/__test__/routes.test.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { createMemoryRouter, RouterProvider } from 'react-router-dom';
 
@@ -5,7 +6,9 @@ import { routes } from '../../routes';
 
 describe('render routes', () => {
   test('test click button to skip page', () => {
-    const router = createMemoryRouter(routes, { initialEntries: ['/vote'] });
+    const router = createMemoryRouter(routes, {
+      initialEntries: ['/proposal'],
+    });
     render(<RouterProvider router={router} />);
     const elements = screen.getAllByTestId('skip-page');
     fireEvent.click(elements[0]);
